Allow same-process execution. Rework eval/include/require runners.
For isolated execution (in sub-processes), the 'prepend' runner now does
everything we could want of it.  Consequently, we no longer need as many
variations on isolated execution.

However, there's still the matter that isolated/sub-process execution makes
debugging a bit more awkward, and we don't yet have a perfect way to execute
within the same process.

This revision shuffles the runners around:

BEFORE: 'eval' and 'include' both launch subprocesses, but they use
different ways to execute the target script in that process.

AFTER: 'eval' and 'require' both launch code in the same process as pogo,
but they use different ways to execute the target script in that process.